1:Refactory Configuration.cs and modify other config element references. 2:Create a build-in plugin \"InternalPlugin\" and merge update plugin
